# Conflicts:
#	gai++/src/DataStore.h
#	gai++/src/DataStoreSqlite.cpp
#	gai++/src/Dispatcher.cpp
#	gai++/src/Hit.cpp
#	gai++/src/Hit.h

+++ gai++/src/DataStore.h
@@ -211,7 +211,7 @@ namespace GAI
         /// @return
         ///     Generated Hit instance
         ///
-        Hit createHit(const int id, const std::string& version, const std::string& url, const double timestamp)
+        Hit createHit(const int id, const std::string& version, const std::string& url, const uint64_t timestamp)
         {
             return Hit(id, version, url, timestamp);
         }
